backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
python-multipart
Pillow
python-dotenv
boto3
